requirements.txt
flask==3.0.3
google-generativeai==0.7.2
python-dotenv==1.0.1
werkzeug==3.0.3
pygments==2.18.0
gunicorn==22.0.0
